youtube chatbot/requirements.txt
langchain==0.2.14
langchain-google-genai
langchain-community
faiss-cpu
youtube-transcript-api
python-dotenv
# streamlit